Create "error_test", move related tests there.
