+++ rjiter/tests/rjiter_test.rs
@@ -3,7 +3,6 @@ use std::sync::Arc;
 
 use jiter::JsonValue;
 use jiter::LazyIndexMap;
-use rjiter::LinePosition;
 use rjiter::NumberInt;
 use rjiter::Peek;
 use rjiter::RJiter;
@@ -468,40 +467,6 @@ fn current_index() {
         let result = rjiter.finish();
         assert!(result.is_ok());
         assert_eq!(rjiter.current_index(), pos_len_done);
-    }
-}
-
-#[test]
-fn index_in_error() {
-    let token_pos = 32;
-    let lot_of_spaces = " ".repeat(token_pos);
-    let input = format!(r#"{lot_of_spaces}"hello""#);
-    let mut buffer = [0u8; 16];
-    let mut reader = Cursor::new(input.as_bytes());
-    let mut rjiter = RJiter::new(&mut reader, &mut buffer);
-
-    let result = rjiter.next_bool();
-    let err = result.unwrap_err();
-    assert_eq!(err.index, token_pos);
-}
-
-#[test]
-fn position_for_error() {
-    let leading_text = "\n \n  \n   \n    \n      \n   ";
-    let input = format!(r#"{leading_text}null null"#);
-    let mut buffer = [0u8; 10];
-    let mut reader = Cursor::new(input.as_bytes());
-    let mut rjiter = RJiter::new(&mut reader, &mut buffer);
-
-    let result = rjiter.next_str();
-    match result {
-        Err(rjiter_err) => {
-            let pos = rjiter_err.get_position(&rjiter);
-            // for column: jiter reports the position of the space after "null"
-            // 3 spaces in `leading_text`, 4 characters in "null"
-            assert_eq!(pos, LinePosition::new(7, 8));
-        }
-        _ => panic!("Expected JiterError"),
     }
 }
